packages/vault-api/requirements-sqlite.txt
# Vault API Dependencies (SQLite version - no PostgreSQL)

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database (SQLite - no psycopg2 needed)
sqlalchemy==2.0.23
alembic==1.12.1

# Redis & Background Jobs (optional)
apscheduler==3.10.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
